requirements.txt
watchdog
pynput
pywin32
python-dotenv
